Rename module `Label` to `label`
