+++ src/asm/mod.rs
@@ -22,4 +22,4 @@ pub mod record;
 pub mod signature;
 mod symbol;
 pub mod types;
-pub mod Label;
+pub mod label;
